requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-allauth==65.14.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
